api_yamdb/requirements.txt
Django==3.2
PyJWT==2.1.0
asgiref==3.3.2
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
djoser==2.1.0
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-return==1.2.0
flake8==5.0.4
gunicorn==20.0.4
pep8-naming==0.13.3
psycopg2-binary==2.8.6
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1